# docs/swagger.yaml
basePath: /api
consumes:
- application/json
definitions:
...
  Artifact:
    description: |-
      The metadata defines properties to associate artifact with a task or job and can be used to query artifacts
      related for a job and an organization.
    properties:
      artifact_order:
        description: ArtifactOrder of artifact in group
        format: int64
        type: integer
        x-go-name: ArtifactOrder
      bucket:
        description: Bucket defines bucket where artifact is stored
        type: string
        x-go-name: Bucket
      content_length:
        description: ContentLength refers to content-length of artifact
        format: int64
        type: integer
        x-go-name: ContentLength
      content_type:
        description: ContentType refers to content-type of artifact
        type: string
        x-go-name: ContentType
      created_at:
        description: CreatedAt job creation time
        format: date-time
        type: string
        x-go-name: CreatedAt
      etag:
        description: ETag stores ETag from underlying storage such as Minio/S3
        type: string
        x-go-name: ETag
      expires_at:
        description: ExpiresAt - expiration time
        format: date-time
        type: string
        x-go-name: ExpiresAt
      group:
        description: Group of artifact
        type: string
        x-go-name: Group
      id:
        description: |-
          gorm.Model
          ID defines UUID for primary key
        type: string
        x-go-name: ID
      job_execution_id:
        description: JobExecutionID refers to job-execution-id being processed
        type: string
        x-go-name: JobExecutionID
      job_request_id:
        description: JobRequestID refers to request-id being processed
        format: uint64
        type: integer
        x-go-name: JobRequestID
      kind:
        description: Kind of artifact
        type: string
        x-go-name: Kind
      metadata:
        additionalProperties:
          type: string
        description: MetadataMap - transient map of properties - deserialized from
          MetadataSerialized
        type: object
        x-go-name: Metadata
      name:
        description: Name defines name of artifact for display
        type: string
        x-go-name: Name
      organization_id:
        description: OrganizationID defines org who submitted the job
        type: string
        x-go-name: OrganizationID
      permissions:
        description: Permissions of artifact
        format: int32
        type: integer
        x-go-name: Permissions
      sha256:
        description: SHA256 defines hash of the contents using SHA-256 algorithm
        type: string
        x-go-name: SHA256
      tags:
        additionalProperties:
          type: string
        type: object
        x-go-name: Tags
      task_execution_id:
        description: TaskExecutionID refers to task-execution-id being processed
        type: string
        x-go-name: TaskExecutionID
      task_type:
        description: TaskType defines type of task
        type: string
        x-go-name: TaskType
      updated_at:
        description: UpdatedAt job update time
        format: date-time
        type: string
        x-go-name: UpdatedAt
      url:
        type: string
        x-go-name: URL
      user_id:
        description: UserID defines user who submitted the job
        type: string
        x-go-name: UserID
    title: Artifact defines metadata of artifact that is uploaded by a job such as
      task logs, task results, etc.
    type: object
    x-go-package: plexobject.com/formicary/internal/types
...
  Duration:
    description: |-
      A Duration represents the elapsed time between two instants
      as an int64 nanosecond count. The representation limits the
      largest representable duration to approximately 290 years.
    format: int64
    type: integer
    x-go-package: time
...
  JobExecution:
    description: |-
      JobExecution refers to a specific instance of a job-definition that gets activated upon the submission of a
      job-request. For every task outlined within the task-definition associated with the JobExecution, a
      corresponding TaskExecution instance is generated. This setup tracks the progress and state of both job and
      task executions within a database, and any outputs generated during the job execution process are preserved in
      object storage.
    properties:
      contexts:
        description: Contexts defines context variables of job
        items:
          $ref: '#/definitions/JobExecutionContext'
        type: array
        x-go-name: Contexts
      cpu_secs:
        description: CPUSecs execution time
        format: int64
        type: integer
        x-go-name: CPUSecs
      ended_at:
        description: EndedAt job execution end time
        format: date-time
        type: string
        x-go-name: EndedAt
      error_code:
        description: ErrorCode captures error code at the end of job execution if
          it fails
        type: string
        x-go-name: ErrorCode
      error_message:
        description: ErrorMessage captures error message at the end of job execution
          if it fails
        type: string
        x-go-name: ErrorMessage
      exit_code:
        description: ExitCode defines exit status from the job execution
        type: string
        x-go-name: ExitCode
      exit_message:
        description: ExitMessage defines exit message from the job execution
        type: string
        x-go-name: ExitMessage
      id:
        description: |-
          gorm.Model
          ID defines UUID for primary key
        type: string
        x-go-name: ID
      job_request_id:
        description: JobRequestID defines foreign key for job request
        format: uint64
        type: integer
        x-go-name: JobRequestID
      job_state:
        $ref: '#/definitions/RequestState'
      job_type:
        description: JobType defines type for the job
        type: string
        x-go-name: JobType
      job_version:
        type: string
        x-go-name: JobVersion
      organization_id:
        description: OrganizationID defines org who submitted the job
        type: string
        x-go-name: OrganizationID
      started_at:
        description: StartedAt job execution start time
        format: date-time
        type: string
        x-go-name: StartedAt
      tasks:
        description: Tasks defines list of tasks that are executed for the job
        items:
          $ref: '#/definitions/TaskExecution'
        type: array
        x-go-name: Tasks
      updated_at:
        description: UpdatedAt job execution last update time
        format: date-time
        type: string
        x-go-name: UpdatedAt
      user_id:
        description: UserID defines user who submitted the job
        type: string
        x-go-name: UserID
    type: object
    x-go-package: plexobject.com/formicary/queen/types
  JobExecutionContext:
    properties:
      created_at:
        description: CreatedAt job context creation time
        format: date-time
        type: string
        x-go-name: CreatedAt
      id:
        description: ID defines UUID for primary key
        type: string
        x-go-name: ID
      job_execution_id:
        description: |-
          gorm.Model
          JobExecutionID defines foreign key for JobExecution
        type: string
        x-go-name: JobExecutionID
      name:
        description: Name defines name of property
        type: string
        x-go-name: Name
      secret:
        description: Secret for encryption
        type: boolean
        x-go-name: Secret
      type:
        description: Type defines type of property value
        type: string
        x-go-name: Type
      value:
        description: Value defines value of property that can be string, number, boolean
          or JSON structure
        type: string
        x-go-name: Value
    title: JobExecutionContext defines context for the job execution.
    type: object
    x-go-package: plexobject.com/formicary/queen/types
...
  JobRequest:
    description: |-
      JobRequest outlines a user's request to execute a job as per its job-definition. Upon submission, a job-request
      is marked as PENDING in the database and later, it is asynchronously scheduled for execution by the job scheduler,
      depending on resource availability. Besides user-initiated requests, a job request might also be issued by a
      parent job to execute a child job in a fork/join manner.
    properties:
      created_at:
        description: CreatedAt job creation time
        format: date-time
        type: string
        x-go-name: CreatedAt
      cron_triggered:
        description: CronTriggered is true if request was triggered by cron
        type: boolean
        x-go-name: CronTriggered
      description:
        description: Description of the request
        type: string
        x-go-name: Description
      error_code:
        description: ErrorCode captures error code at the end of job execution if
          it fails
        type: string
        x-go-name: ErrorCode
      error_message:
        description: ErrorMessage captures error message at the end of job execution
          if it fails
        type: string
        x-go-name: ErrorMessage
      execution:
        $ref: '#/definitions/JobExecution'
      id:
        description: |-
          gorm.Model
          ID defines UUID for primary key
        format: uint64
        type: integer
        x-go-name: ID
      job_definition_id:
        description: JobDefinitionID points to the job-definition version
        type: string
        x-go-name: JobDefinitionID
      job_execution_id:
        description: JobExecutionID defines foreign key for JobExecution
        type: string
        x-go-name: JobExecutionID
      job_group:
        description: JobGroup defines a property for grouping related job
        type: string
        x-go-name: JobGroup
      job_priority:
        description: JobPriority defines priority of the job
        format: int64
        type: integer
        x-go-name: JobPriority
      job_state:
        $ref: '#/definitions/RequestState'
      job_type:
        description: JobType defines type for the job
        type: string
        x-go-name: JobType
      job_version:
        type: string
        x-go-name: JobVersion
      last_job_execution_id:
        description: LastJobExecutionID defines foreign key for JobExecution
        type: string
        x-go-name: LastJobExecutionID
      organization_id:
        description: OrganizationID defines org who submitted the job
        type: string
        x-go-name: OrganizationID
      params:
        additionalProperties:
          type: object
        type: object
        x-go-name: NameValueParams
      parent_id:
        description: ParentID defines id for parent job
        format: uint64
        type: integer
        x-go-name: ParentID
      permissions:
        description: Permissions provides who can access this request 0 - all, 1 -
          Org must match, 2 - UserID must match from authentication
        format: int64
        type: integer
        x-go-name: Permissions
      platform:
        description: Platform overrides platform property for targeting job to a specific
          follower
        type: string
        x-go-name: Platform
      quick_search:
        description: QuickSearch provides quick search to search a request by params
        type: string
        x-go-name: QuickSearch
      retried:
        description: Retried keeps track of retry attempts
        format: int64
        type: integer
        x-go-name: Retried
      schedule_attempts:
        description: ScheduleAttempts defines attempts of schedule
        format: int64
        type: integer
        x-go-name: ScheduleAttempts
      scheduled_at:
        description: |-
          ScheduledAt defines schedule time when job will be submitted so that you can submit a job
          that will be executed later
        format: date-time
        type: string
        x-go-name: ScheduledAt
      timeout:
        $ref: '#/definitions/Duration'
      updated_at:
        description: UpdatedAt job update time
        format: date-time
        type: string
        x-go-name: UpdatedAt
      user_id:
        description: UserID defines user who submitted the job
        type: string
        x-go-name: UserID
      user_key:
        description: UserKey defines user-defined UUID and can be used to detect duplicate
          jobs
        type: string
        x-go-name: UserKey
    type: object
    x-go-package: plexobject.com/formicary/queen/types
...
  RequestState:
    title: RequestState defines enum for the state of request or execution such as
      pending/completed/failed.
    type: string
    x-go-package: plexobject.com/formicary/internal/types
...
  TaskExecution:
    description: |-
      TaskExecution records the execution of a task or a unit of work, carried out by ant-workers in accordance
      with the specifications of the task-definition. It captures the status and the outputs produced by the
      task execution, storing them in the database and the object-store.
    properties:
      allow_failure:
        description: AllowFailure means the task is optional and can fail without
          failing entire job
        type: boolean
        x-go-name: AllowFailure
      ant_host:
        description: AntHost - host where ant ran the task
        type: string
        x-go-name: AntHost
      ant_id:
        description: AntID - id of ant with version
        type: string
        x-go-name: AntID
      artifacts:
        description: Artifacts defines list of artifacts that are generated for the
          task
        items:
          $ref: '#/definitions/Artifact'
        type: array
        x-go-name: Artifacts
      contexts:
        description: Contexts defines context variables of task
        items:
          $ref: '#/definitions/TaskExecutionContext'
        type: array
        x-go-name: Contexts
      cost_factor:
        description: CostFactor
        format: double
        type: number
        x-go-name: CostFactor
      count_services:
        description: CountServices
        format: int64
        type: integer
        x-go-name: CountServices
      ended_at:
        description: EndedAt job update time
        format: date-time
        type: string
        x-go-name: EndedAt
      error_code:
        description: ErrorCode captures error code at the end of job execution if
          it fails
        type: string
        x-go-name: ErrorCode
      error_message:
        description: ErrorMessage captures error message at the end of job execution
          if it fails
        type: string
        x-go-name: ErrorMessage
      exit_code:
        description: ExitCode defines exit status from the job execution
        type: string
        x-go-name: ExitCode
      exit_message:
        description: ExitMessage defines exit message from the job execution
        type: string
        x-go-name: ExitMessage
      failed_command:
        description: FailedCommand captures command that failed
        type: string
        x-go-name: FailedCommand
      id:
        description: |-
          gorm.Model
          ID defines UUID for primary key
        type: string
        x-go-name: ID
      job_execution_id:
        description: JobExecutionID defines foreign key for JobExecution
        type: string
        x-go-name: JobExecutionID
      method:
        $ref: '#/definitions/TaskMethod'
      retried:
        description: Retried keeps track of retry attempts
        format: int64
        type: integer
        x-go-name: Retried
      started_at:
        description: StartedAt job creation time
        format: date-time
        type: string
        x-go-name: StartedAt
      stdout:
        items:
          type: string
        type: array
        x-go-name: Stdout
      task_order:
        description: TaskOrder
        format: int64
        type: integer
        x-go-name: TaskOrder
      task_state:
        $ref: '#/definitions/RequestState'
      task_type:
        description: TaskType defines type of task
        type: string
        x-go-name: TaskType
      updated_at:
        description: UpdatedAt job execution last update time
        format: date-time
        type: string
        x-go-name: UpdatedAt
    type: object
    x-go-package: plexobject.com/formicary/queen/types
  TaskExecutionContext:
    properties:
      created_at:
        description: CreatedAt task context creation time
        format: date-time
        type: string
        x-go-name: CreatedAt
      id:
        description: ID defines UUID for primary key
        type: string
        x-go-name: ID
      name:
        description: Name defines name of property
        type: string
        x-go-name: Name
      secret:
        description: Secret for encryption
        type: boolean
        x-go-name: Secret
      task_execution_id:
        description: TaskExecutionID defines foreign key for task-execution
        type: string
        x-go-name: TaskExecutionID
      type:
        description: Type defines type of property value
        type: string
        x-go-name: Type
      value:
        description: Value defines value of property that can be string, number, boolean
          or JSON structure
        type: string
        x-go-name: Value
    title: TaskExecutionContext defines context variables for the task execution that
      are captured by the task executor.
    type: object
    x-go-package: plexobject.com/formicary/queen/types
  TaskMethod:
    description: |-
      The ant followers registers with the methods that they support and the task is then routed
      based on method, tags and concurrency of the ant follower.
    title: TaskMethod defines enum for method of task when communicating with remote
      ant follower.
    type: string
    x-go-package: plexobject.com/formicary/internal/types
...
host: https://formicary.io
info:
  contact:
    email: support@formicary.io
    name: Support
    url: https://formicary.io
  description: |-
    The formicary is a distributed orchestration engine based on `Leader-Follower`, `Pipes-Filter`, `Fork-Join` and `SEDA` design principles for
    executing a directed acyclic graph of tasks, which is also referred as a job workflow. A task represents a unit of work and a job definition is used to specify the task
    dependencies in the graph/workflow including configuration parameters and conditional logic.
  license:
    name: AGPL
    url: https://opensource.org/licenses/AGPL-3.0
  title: Formicary API.
  version: 0.0.1
paths:
...
  /api/jobs/requests:
    get:
      operationId: queryJobRequests
      parameters:
      - format: int64
        in: query
        name: page
        type: integer
        x-go-name: Page
      - format: int64
        in: query
        name: page_size
        type: integer
        x-go-name: PageSize
      - description: JobType defines a unique type of job
        in: query
        name: job_type
        type: string
        x-go-name: JobType
      - description: Platform can be OS platform or target runtime and a job can be
          targeted for specific platform that can be used for filtering
        in: query
        name: platform
        type: string
        x-go-name: Platform
      - description: JobState defines state of job that is maintained throughout the
          lifecycle of a job
        in: query
        name: job_state
        type: string
        x-go-name: JobState
      - description: JobGroup defines a property for grouping related job
        in: query
        name: job_group
        type: string
        x-go-name: JobGroup
      responses:
        "200":
          $ref: '#/responses/jobRequestQueryResponse'
      summary: Queries job requests by criteria such as type, platform, etc.
      tags:
      - job-requests
    post:
      operationId: submitJobRequest
      parameters:
      - in: body
        name: Body
        schema:
          $ref: '#/definitions/JobRequest'
      responses:
        "200":
          $ref: '#/responses/jobRequest'
      summary: Submits a job-request for processing, which is saved in the database
        and is then scheduled for execution.
      tags:
      - job-requests
  /api/jobs/requests/{id}:
    get:
      operationId: getJobRequest
      parameters:
      - in: path
        name: id
        required: true
        type: string
        x-go-name: ID
      responses:
        "200":
          $ref: '#/responses/jobRequest'
      summary: Finds the job-request by id.
      tags:
      - job-requests
  /api/jobs/requests/{id}/cancel:
    post:
      operationId: cancelJobRequest
      parameters:
      - in: path
        name: id
        required: true
        type: string
        x-go-name: ID
      responses:
        "200":
          $ref: '#/responses/emptyResponse'
      summary: Cancels a job-request that is pending for execution or already executing.
      tags:
      - job-requests
  /api/jobs/requests/{id}/dot:
    get:
      operationId: dotJobRequest
      parameters:
      - in: path
        name: id
        required: true
        type: string
        x-go-name: ID
      responses:
        "200":
          $ref: '#/responses/stringResponse'
      summary: Returns Graphviz DOT request for the graph of tasks defined in the
        job request.
      tags:
      - job-requests
  /api/jobs/requests/{id}/dot.png:
    get:
      operationId: dotImageJobRequest
      parameters:
      - in: path
        name: id
        required: true
        type: string
        x-go-name: ID
      responses:
        "200":
          $ref: '#/responses/byteResponse'
      summary: Returns Graphviz DOT image for the graph of tasks defined in the job.
      tags:
      - job-requests
  /api/jobs/requests/{id}/restart:
    post:
      operationId: restartJobRequest
      parameters:
      - in: path
        name: id
        required: true
        type: string
        x-go-name: ID
      responses:
        "200":
          $ref: '#/responses/emptyResponse'
      summary: Restarts a previously failed job so that it can re-execute, the restart
        may perform soft-restart where only failed tasks are executed or hard-restart
        where all tasks are executed.
      tags:
      - job-requests
  /api/jobs/requests/{id}/trigger:
    post:
      operationId: triggerJobRequest
      parameters:
      - in: path
        name: id
        required: true
        type: string
        x-go-name: ID
      responses:
        "200":
          $ref: '#/responses/emptyResponse'
      summary: Triggers a scheduled job.
      tags:
      - job-requests
...
produces:
- application/json
responses:
...
  byteResponse:
    description: Byte Array response body
    schema:
      items:
        format: uint8
        type: integer
      type: array
...
  emptyResponse:
    description: Empty response body
...
  jobRequest:
    description: JobRequest defines user request to process a job, which is saved
      in the database as PENDING and is then scheduled for job execution.
    schema:
      $ref: '#/definitions/JobRequest'
...
  jobRequestQueryResponse:
    description: Paginated results of jobRequests matching query
    schema:
      properties:
        Page:
          format: int64
          type: integer
        PageSize:
          format: int64
          type: integer
        Records:
          items:
            $ref: '#/definitions/JobRequest'
          type: array
        TotalPages:
          format: int64
          type: integer
        TotalRecords:
          format: int64
          type: integer
      type: object
...
  stringResponse:
    description: String response body
...
schemes:
- http
- https
security:
- api_key: []
securityDefinitions:
  api_key:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"
...